refactor: move highest damage punish to typescript
