+++ src/lib/stats/definitions/index.js
@@ -11,6 +11,7 @@ export * from "./lCancelAccuracy";
 export * from "./neutralOpenerMoves";
 export * from "./killMoves";
 export * from "./selfDestructs";
+export * from "./highestDamagePunish";
 
 export const openingsPerKill = {
   name: "Openings / Kill",
@@ -145,41 +146,6 @@ export const averageKillPercent = {
     return {
       result: result,
       simple: genSimpleFromRatio(result, this.recommendedRounding),
-    };
-  },
-};
-
-export const highDamagePunishes = {
-  name: "Highest Damage Punish",
-  type: "number",
-  betterDirection: "higher",
-  recommendedRounding: 1,
-  calculate(games, playerIndex) {
-    const punishes = _.flatMap(games, (game) => {
-      const conversions = _.get(game, ["stats", "conversions"]) || [];
-      return _.filter(conversions, (conversion) => {
-        const isForPlayer = conversion.playerIndex === playerIndex;
-        const hasEndPercent = conversion.endPercent !== null;
-        return isForPlayer && hasEndPercent;
-      });
-    });
-
-    const getDamageDone = (punish) => punish.endPercent - punish.startPercent;
-    const orderedPunishes = _.orderBy(punishes, [getDamageDone], "desc");
-    const topPunish = _.first(orderedPunishes);
-    const simple = {
-      text: "N/A",
-      number: null,
-    };
-
-    if (topPunish) {
-      simple.number = getDamageDone(topPunish);
-      simple.text = simple.number.toFixed(this.recommendedRounding);
-    }
-
-    return {
-      result: _.take(orderedPunishes, 5),
-      simple: simple,
     };
   },
 };
